fix: Following 멤버의 글 목록에서 닉네임,아바타url 안보이는 이슈 수정

+++ src/main/java/click/dailyfeed/timeline/domain/post/mapper/TimelinePostMapper.java
@@ -26,7 +26,7 @@ public interface TimelinePostMapper {
                 .build();
     }
 
-    default PostDto.Post toPostDto(Post post, MemberProfileDto.Summary author, Integer commentCount) {
+    default PostDto.Post toPostDto(Post post, MemberProfileDto.Summary author, Long commentCount) {
         return PostDto.Post.builder()
                 .id(post.getId())
                 .title(post.getTitle())
